restcontroller fixed

+++ src/main/java/com/example/SingleTaskManagement/repository/TaskRepository.java
@@ -5,8 +5,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TaskRepository extends JpaRepository<Task, Long> {
+public interface TaskRepository extends JpaRepository<Task, Integer> {
     List<Task> findAll();
+
+   
+  
+
     List<Task> findById(long id);
+
 
 }
